+++ nv/vim.py
@@ -79,6 +79,9 @@ DIRECTION_DOWN = 2
 DIRECTION_LEFT = 3
 DIRECTION_RIGHT = 4
 
+EOF = '\x00'
+NL = '\n'
+
 
 def console_message(msg):
     # type: (str) -> None